rename task_param_reader as stim_param_reader
