+++ examples/configs/tasks/ahhh_task_1.yml
@@ -2,7 +2,7 @@ stimulus_id: ahhh_task_1
 stimulus_description: Speech ahh
 num_iterations: 1
 duration: 10.0
-arg_parser: iout.task_param_reader.py::SpeechStimArgs()
+arg_parser: iout.stim_param_reader.py::SpeechStimArgs()
 stimulus_file_type: psychopy
 stimulus_file: tasks.task.py::Task_countdown()
 countdown: countdown_speech_2022_04_11.mp4
