+++ examples/configs/tasks/coord_pause_task_2.yml
@@ -2,7 +2,7 @@ stimulus_id: coord_pause_task_2
 stimulus_description: 'pause'
 num_iterations: 1
 duration: null
-arg_parser: iout.task_param_reader.py::CoordPause2StimArgs
+arg_parser: iout.stim_param_reader.py::CoordPause2StimArgs
 stimulus_file_type: psychopy
 stimulus_file: tasks.task.py::Task_pause()
 slide_image: please_wait_slide.jpg
