+++ examples/configs/tasks/finger_nose_demo_task_1.yml
@@ -2,7 +2,7 @@ stimulus_id: finger_nose_demo_task_1
 stimulus_description: finger to nose
 num_iterations: 1
 duration: 10.0
-arg_parser: iout.task_param_reader.py::FingerNoseStimArgs()
+arg_parser: iout.stim_param_reader.py::FingerNoseStimArgs()
 stimulus_file_type: psychopy
 stimulus_file: tasks.fixations.py::Fixation_Target_sidetrials()
 target_pos:
