+++ examples/configs/tasks/pursuit_task_1.yml
@@ -2,7 +2,7 @@ stimulus_id: pursuit_task_1
 stimulus_description: pursuit circle
 num_iterations: 2
 duration: null
-arg_parser: iout.task_param_reader.py::PursuitStimArgs()
+arg_parser: iout.stim_param_reader.py::PursuitStimArgs()
 stimulus_file_type: psychopy
 stimulus_file: tasks.smooth_pursuit.pursuit_task.py::Pursuit()
 target_size: 0.7
