+++ examples/configs/tasks/saccades_vertical_task_1.yml
@@ -2,7 +2,7 @@ stimulus_id: saccades_vertical_task_1
 stimulus_description: vertical sacade
 num_iterations: 1
 duration: null
-arg_parser: iout.task_param_reader.py::SaccadesStimArgs()
+arg_parser: iout.stim_param_reader.py::SaccadesStimArgs()
 stimulus_file_type: psychopy
 stimulus_file: tasks.saccade.saccade_task.py::Saccade()
 direction: vertical
