+++ examples/configs/tasks/sit_to_stand_task_1.yml
@@ -2,6 +2,6 @@ stimulus_id: sit_to_stand_task_1
 stimulus_description: Sit to Stand
 num_iterations: 5
 duration: null
-arg_parser: iout.task_param_reader.py::EyeTrackerStimArgs()
+arg_parser: iout.stim_param_reader.py::EyeTrackerStimArgs()
 stimulus_file_type: psychopy
 stimulus_file: tasks.task.py::Task()
